Fix: 실적 목표 등록 기능 테스트 중 발생한 오류 수정
원활한 테스트를 위한 더미 데이터 수정

+++ src/main/java/com/clover/salad/goal/command/domain/aggregate/entity/Goal.java
@@ -38,7 +38,7 @@ public class Goal {
 	private Long totalRentalAmount;
 	
 	@Column(name = "customer_feedback_score")
-	private Integer customerFeedbackScore;
+	private Double customerFeedbackScore;
 	
 	@Column(name = "customer_feedback_count")
 	private Integer customerFeedbackCount;
